Add MenuBuilderService Add SaveEntityService
